Switch CORS proxies yet again.
allorigins.me is dead.
iplayif's proxy has some hope of being stable, since it's maintained by
a real person for a specific purpose (playing IF games in Parchment).

Also, continue factoring code into `cors_proxy` as its purpose solidifies.
The idea is that we should be able to give it a regular URL and a callback,
and it should call that callback with the text retrieved from that URL
no matter what proxy-specific stuff we need to do to wrap and unwrap it.